void list_distroy(struct node **p_list);
void list_destroy(struct node **p_list)
{
    if ((*p_list)->next)
        list_destroy(&((*p_list)->next));
    free(*p_list);
}

Frees all memory allocated for the list.
If there are any references to the list, this function doesn't care! Be
careful - you could cause a seg fault.
If you create a list within a subroutine, then use this at the end if
you are not going to use the list again.